Tomris Bot/requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
certifi==2024.2.2
pytz==2023.3
